back/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
psycopg[binary]==3.2.3
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
